+++ src/RoslynAnalyzers/Utilities/Compiler/Extensions/PooledHashSetExtensions.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Collections.Generic;
+using System.Collections.Immutable;
 using Microsoft.CodeAnalysis.PooledObjects;
 
 namespace Analyzer.Utilities.PooledObjects.Extensions
@@ -16,5 +17,31 @@ namespace Analyzer.Utilities.PooledObjects.Extensions
                 builder.Add(item);
             }
         }
+    }
+}
+
+namespace Microsoft.CodeAnalysis.PooledObjects
+{
+    internal sealed partial class PooledHashSet<T>
+    {
+        public ImmutableHashSet<T> ToImmutableAndFree()
+        {
+            ImmutableHashSet<T> result;
+            if (Count == 0)
+            {
+                result = ImmutableHashSet<T>.Empty;
+            }
+            else
+            {
+                result = this.ToImmutableHashSet(Comparer);
+                this.Clear();
+            }
+
+            _pool?.Free(this);
+            return result;
+        }
+
+        public ImmutableHashSet<T> ToImmutable()
+            => Count == 0 ? ImmutableHashSet<T>.Empty : this.ToImmutableHashSet(Comparer);
     }
 }
