+++ src/RoslynAnalyzers/Utilities/Compiler/PooledObjects/PooledDictionary.cs
@@ -59,27 +59,4 @@ namespace Microsoft.CodeAnalysis.PooledObjects
             return instance;
         }
     }
-
-    internal sealed partial class PooledHashSet<T>
-    {
-        public ImmutableHashSet<T> ToImmutableAndFree()
-        {
-            ImmutableHashSet<T> result;
-            if (Count == 0)
-            {
-                result = ImmutableHashSet<T>.Empty;
-            }
-            else
-            {
-                result = this.ToImmutableHashSet(Comparer);
-                this.Clear();
-            }
-
-            _pool?.Free(this);
-            return result;
-        }
-
-        public ImmutableHashSet<T> ToImmutable()
-            => Count == 0 ? ImmutableHashSet<T>.Empty : this.ToImmutableHashSet(Comparer);
-    }
 }
